Reuse theme in docs
